+++ block/src/formats/vhdx/internal/header.rs
@@ -341,7 +341,7 @@ impl RegionTableEntry {
         // SAFETY: the assertion above makes sure the buffer size is correct.
         let mut region_table_entry: RegionTableEntry = unsafe { *(buffer.as_ptr().cast()) };
 
-        let uuid = crate::vhdx::uuid_from_guid(buffer);
+        let uuid = super::uuid_from_guid(buffer);
         region_table_entry.guid = uuid;
 
         Ok(region_table_entry)
