+++ block/src/formats/vhdx/worker/mod.rs
@@ -1,0 +1,7 @@
+// Copyright 2026 The Cloud Hypervisor Authors. All rights reserved.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+//! Sync I/O worker for VHDX images.
+
+pub(crate) mod sync;
